requirements.txt
requests
openai
duckduckgo-search
fastapi
uvicorn
python-dotenv
